refactor(task_executor/pr_detection): extract tests via #[path]
Phase 0 PR-9 of the architecture decoupling roadmap. Fourth clean
U-16 victory: extraction brings pr_detection.rs fully under the
800-line ceiling.

Net effect:
  task_executor/pr_detection.rs:        1350 -> 728 lines (-622)  ← under U-16 ✓
  task_executor/pr_detection_tests.rs:   NEW   622 lines

Verification:
- cargo test -p harness-server --lib -- task_executor::pr_detection:
    48/48 passed (covers all pr_detection::tests::* including
    existing_pr_lookup_follows_next_page,
    existing_pr_lookup_stops_after_page_limit_without_error)
- RUSTFLAGS=-Dwarnings cargo check --workspace --all-targets: clean
- cargo fmt --all -- --check: clean
